Break up Interface.php for PSR-4 compliance
Each interface needs to be in a file of the same name.
